requirements.txt
databricks-connect==17.3.5
databricks-sdk==0.91.0
sqlglot==28.10.1
scikit-learn==1.8.0
plotly==6.5.2
pandas==2.3.3
numpy==2.4.2
python-dotenv==1.2.1
networkx